+++ src/Method/WebAuth.php
@@ -7,15 +7,19 @@ namespace Yiisoft\User\Method;
 use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
-use Yiisoft\Auth\AuthenticatorWithChallengeInterface;
+use Yiisoft\Auth\AuthenticationMethodInterface;
 use Yiisoft\Auth\IdentityInterface;
 use Yiisoft\Http\Status;
 use Yiisoft\User\CurrentUser;
 
 /**
  * Implementation of the `AuthenticatorWithChallengeInterface` for authenticating users in the web applications.
+ *
+ * @psalm-suppress DeprecatedInterface Implements the deprecated `AuthenticationMethodInterface` (instead of
+ * `AuthenticatorWithChallengeInterface` directly) to avoid a backward compatibility break. Will be switched
+ * to `AuthenticatorWithChallengeInterface` in the next major version, see https://github.com/yiisoft/user/issues/130.
  */
-final class WebAuth implements AuthenticatorWithChallengeInterface
+final class WebAuth implements AuthenticationMethodInterface
 {
     private string $authUrl = '/login';