+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 2.3.3 under development
 
 - Enh #120: Explicitly import classes and constants in "use" section (@vjik)
-- Enh #127: Bump `yiisoft/auth` version to `^3.3.0`, and fix deprecated classes usage (@klsoft-web)
+- Enh #127: Bump `yiisoft/auth` version to `^3.3.0`, and fix deprecated classes usage (@klsoft-web, @vjik)
 
 ## 2.3.2 December 23, 2025